get pulls flower names, styled dropdown
able to get request to pull only the flower names. Styled the dropdown. Next, allow one of the options to be seleted as the current, and pull images from api to be displayed. Then, apply same logic to other options. Then, fix dropdown style so it overlays ontop of everything and other arrows dont show